fix: clean up product content section editor
